feat: 파일이나 폴더 CRUD시 권한 체크 하도록 변경

+++ src/main/java/com/seohamin/fshs/v2/domain/file/service/FileService.java
@@ -8,7 +8,10 @@ import com.seohamin.fshs.v2.domain.file.dto.FileUploadResponseDto;
 import com.seohamin.fshs.v2.domain.file.entity.File;
 import com.seohamin.fshs.v2.domain.file.entity.Status;
 import com.seohamin.fshs.v2.domain.file.repository.FileRepository;
+import com.seohamin.fshs.v2.domain.folder.entity.Folder;
 import com.seohamin.fshs.v2.domain.folder.repository.FolderRepository;
+import com.seohamin.fshs.v2.domain.user.entity.User;
+import com.seohamin.fshs.v2.domain.user.repository.UserRepository;
 import com.seohamin.fshs.v2.global.exception.CustomException;
 import com.seohamin.fshs.v2.global.exception.constants.ExceptionCode;
 import com.seohamin.fshs.v2.global.infra.ffmpeg.FfmpegProcessor;
@@ -32,6 +35,7 @@ public class FileService {
 
     private final FileRepository fileRepository;
     private final FolderRepository folderRepository;
+    private final UserRepository userRepository;
     private final StorageManager storageManager;
     private final FileUploadProcessor fileUploadProcessor;
     private final Cache<Long, String> filePathCache;
@@ -48,32 +52,36 @@ public class FileService {
     public FileUploadResponseDto uploadFile(
             final MultipartFile multipartFile,
             final Instant lastModified,
-            final Long folderId
+            final Long folderId,
+            final String username
     ) {
 
         // 1) null 검사
-        if (multipartFile == null || lastModified == null || folderId == null) {
+        if (multipartFile == null || lastModified == null || folderId == null || username == null) {
             throw new CustomException(ExceptionCode.INVALID_REQUEST);
         }
 
-        // 2) 시스템 루트 검사
-        if (folderId == 1L) {
-            throw new CustomException(ExceptionCode.SYSTEM_ROOT_FORBIDDEN);
+        // 2) 유저 조회
+        final User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_EXIST));
+
+        // 3) 상위 폴더 조회
+        final Folder folder = folderRepository.findById(folderId)
+                .orElseThrow(() -> new CustomException(ExceptionCode.FOLDER_NOT_EXIST));
+
+        // 4) 업로드 폴더가 해당 유저의 루트 폴더 하위인지 판단
+        if (!hasPermission(user, folder)) {
+            throw new CustomException(ExceptionCode.INVALID_PATH);
         }
 
-        // 3) 상위 폴더 존재 확인 - 잘못된 folderId는 동기 단계에서 즉시 거절
-        if (!folderRepository.existsById(folderId)) {
-            throw new CustomException(ExceptionCode.FOLDER_NOT_EXIST);
-        }
-
-        // 4) 임시 폴더에 파일 저장 - MultipartFile 수명 문제 때문에 요청 스레드에서 동기로 처리
+        // 5) 임시 폴더에 파일 저장 - MultipartFile 수명 문제 때문에 요청 스레드에서 동기로 처리
         final Path tempFilePath = storageManager.saveTemporarily(multipartFile);
 
-        // 5) 파일 UUID 발급 및 처리 중 상태 등록
+        // 6) 파일 UUID 발급 및 처리 중 상태 등록
         final String fileUuid = UUID.randomUUID().toString();
         fileStatusCache.put(fileUuid, Status.PROCESSING);
 
-        // 6) 비동기 처리 위임
+        // 7) 비동기 처리 위임
         fileUploadProcessor.process(fileUuid, tempFilePath, folderId, lastModified);
 
         return new FileUploadResponseDto(fileUuid);
@@ -116,7 +124,10 @@ public class FileService {
      * @param fileId 조회할 파일 ID
      * @return 파일 정보 담긴 DTO
      */
-    public FileResponseDto getFileDetails(final Long fileId) {
+    public FileResponseDto getFileDetails(
+            final Long fileId,
+            final String username
+    ) {
         // 1) null 검사
         if(fileId == null) {
             throw new CustomException(ExceptionCode.INVALID_REQUEST);
@@ -126,7 +137,16 @@ public class FileService {
         final File file = fileRepository.findById(fileId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.FILE_NOT_EXIST));
 
-        // 3) DTO에 담기
+        // 3) 유저 조회
+        final User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_EXIST));
+
+        // 4) 접근 권한 확인
+        if (!hasPermission(user, file)) {
+            throw new CustomException(ExceptionCode.INVALID_PATH);
+        }
+
+        // 5) DTO에 담기
         return FileResponseDto.of(file);
     }
 
@@ -136,7 +156,10 @@ public class FileService {
      * @param fileId 프론트로 전송할 파일의 ID
      * @return 파일의 정보가 담긴 DTO
      */
-    public FileDownloadResponseDto getFile(final Long fileId) {
+    public FileDownloadResponseDto getFile(
+            final Long fileId,
+            final String username
+    ) {
         // 1) null 검사
         if (fileId == null) {
             throw new CustomException(ExceptionCode.INVALID_REQUEST);
@@ -146,16 +169,25 @@ public class FileService {
         final File file = fileRepository.findById(fileId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.FILE_NOT_EXIST));
 
-        // 3) 단순 정보 추출
+        // 3) 유저 조회
+        final User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_EXIST));
+
+        // 4) 접근 권환 조회
+        if (!hasPermission(user, file)) {
+            throw new CustomException(ExceptionCode.INVALID_PATH);
+        }
+
+        // 5) 단순 정보 추출
         final String name = file.getName();
         final String mimeType = file.getMimeType();
         final Long size = file.getSize();
         final String relativePath = file.getRelativePath();
 
-        // 4) 파일 읽어오기
+        // 6) 파일 읽어오기
         final Resource resource = storageManager.getFile(Path.of(relativePath));
 
-        // 5) DTO 조립
+        // 7) DTO 조립
         return new FileDownloadResponseDto(
                 name,
                 mimeType,
@@ -199,7 +231,10 @@ public class FileService {
      * @param fileId 휴지통으로 보낼 메서드
      */
     @Transactional
-    public void deleteFile(final Long fileId) {
+    public void deleteFile(
+            final Long fileId,
+            final String username
+    ) {
         // 1) null 검사
         if (fileId == null) {
             throw new CustomException(ExceptionCode.INVALID_REQUEST);
@@ -209,11 +244,48 @@ public class FileService {
         final File file = fileRepository.findById(fileId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.FILE_NOT_EXIST));
 
-        // 3) DB에서 파일 삭제
+        // 3) 유저 조회
+        final User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_EXIST));
+
+        // 4) 접근 권환 조회
+        if (!hasPermission(user, file)) {
+            throw new CustomException(ExceptionCode.INVALID_PATH);
+        }
+
+        // 5) DB에서 파일 삭제
         fileRepository.delete(file);
 
-        // 4) 실제 파일 삭제
+        // 6) 실제 파일 삭제
         storageManager.removeFile(file.getRelativePath());
+    }
+
+    /**
+     * 해당 유저가 폴더에 접근 권한을 가졌는지 확인하는 메서드
+     * @param user 접근 권한 확인할 유저
+     * @param folder 접근 권한 확인할 폴더
+     * @return 접근 가능 여부
+     */
+    private boolean hasPermission(final User user, final Folder folder) {
+        final String userRootPath = user.getRootFolder().getRelativePath();
+        if (userRootPath.isEmpty()) return true; // 유저 루트 폴더가 시스템 루트 폴더일 때 예외 처리
+        final Path userRootFolderPath = Path.of(userRootPath).normalize();
+        final Path folderPath = Path.of(folder.getRelativePath()).normalize();
+        return folderPath.startsWith(userRootFolderPath);
+    }
+
+    /**
+     * 해당 유저가 파일에 접근 권한을 가졌는지 확인하는 메서드
+     * @param user 접근 권한 확인할 유저
+     * @param file 접근 권한 확인할 파일
+     * @return 접근 가능 여부
+     */
+    private boolean hasPermission(final User user, final File file) {
+        final String userRootPath = user.getRootFolder().getRelativePath();
+        if (userRootPath.isEmpty()) return true; // 유저 루트 폴더가 시스템 루트 폴더일 때 예외 처리
+        final Path userRootFolderPath = Path.of(userRootPath).normalize();
+        final Path folderPath = Path.of(file.getRelativePath()).normalize();
+        return folderPath.startsWith(userRootFolderPath);
     }
 
 }
